+++ src/extension.ts
@@ -30,8 +30,10 @@ export function activate(context: vscode.ExtensionContext): void {
 
     vscode.workspace.onDidChangeTextDocument(changeEvent => {
         let change: vscode.TextDocumentContentChangeEvent = changeEvent.contentChanges[0];
+
+        let changeTokens: string[] = change.text.split(" ");
         /* we insert documentation when a DocMachine comment is created */
-        if (change.text.split(" ")[0] === "/**\n") {
+        if (changeTokens[0] === "/**\n") {
             const editor: vscode.TextEditor | undefined = vscode.window.activeTextEditor;
             let document: vscode.TextDocument = changeEvent.document;
             /* check that the document being edited is the document with changes */