+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## [1.0.7]
+### Changed
+- Fixed bug where the tokens before the function return type would be included as part of the return type
+- Fixed bug where a declaration with a single pointer parameter would result in incorrect return type and parameter names
+
+## [1.0.6]
+### Changed
+- Fixed changelog
+- Fixed bug where the function name would be inserted as a parameter name when the function had no parameters
+- Fixed issues caused by patch 1.0.4
+
 ## [1.0.5]
 - New square icon
 - Fixed bug where the function name would be inserted as a parameter name when the function had no parameters